% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoordBase.R
\name{Extract.coords}
\alias{Extract.coords}
\alias{[.coords}
\title{Extract or Replace Parts of a Coords Object}
\usage{
\method{[}{coords}(x, i)
}
\arguments{
\item{x}{a \code{"coords"} object.}

\item{i}{indices specifying elements to extract or replace—see \code{\link[base:Extract]{Extract}}.}
}
\value{
a \code{"coords"} object.
}
\description{
Extract or replace subsets of coords.
}
\details{
The index may not reference values greater than \code{length(x)} without throwing an error.
}
\examples{
## Continuing example from `as_coords()`...
\dontshow{
   dm <-
       c(5130.4659, 4932.7726, 4806.4339, 3853.3696, 0.0000, -3706.7044, -5306.2869, -2514.4093,
       -007.6754, 1823.9137, -12246.7203, -7702.1145, 0.0000, -1217.3178, 7331.0370, -5731.1536)
   names(dm) <- 
       rep(c("Nelson's Column", "Ostravice", "Tally Ho", "Washington Monument", "Null Island",
             "Tristan da Cunha", "Mawson Peak", "Silvio Pettirossi International Airport"), 2)
   invisible(as_coords(dm, fmt = 2))
   latlon(dm) <- rep(c(TRUE, FALSE), each = 8)
}

## Named "coords" object in degrees and minutes with
## eight values each of latitude and longitude
dm

dm[1:8]

dm[-8:0]

rm(dm)
}
\seealso{
\code{"\link{coords}"}, \code{\link[base:Extract]{Extract}}.
}
\concept{extract}
