% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{latlon-set}
\alias{latlon-set}
\alias{latlon<-}
\title{Latitude or Longitude Attribute for Coords}
\usage{
latlon(nv) <- value
}
\arguments{
\item{nv}{object of class \code{\link[=coords]{"coords"}}.}

\item{value}{a \code{boolean vector} of length \code{1L} or of \code{length(nv)}.}
}
\value{
Argument \code{nv} is returned with \code{boolean vector} attribute \code{"latlon"}
updated as appropriate.
}
\description{
\code{latlon()<-} adds the attribute \code{"latlon"} to objects of class
\code{\link[=coords]{"coords"}}, or modifies an existing \code{"latlon"} attribute.
}
\details{
Attribute \code{"latlon"} is a \code{boolean vector} of length \code{1L} or \code{length(nv)}
for which \code{TRUE} values represent latitude and \code{FALSE} values represent longitude.
Setting this attribute to any other length will result in an error. A \code{boolean vector} of
length \code{1L} signifies that values are all latitude if \code{TRUE}, or longitude if
\code{false}.

This attribute is used in formatting printed output and also by
\code{\link[=validate]{validate()}}. Indeed, the values of \code{nv} are revalidated every time
attribute \code{"latlon"} is added or changed.
}
