\name{Waypoint-package}
\alias{Waypoint-package}
\alias{Waypoint}
\docType{package}
\title{\packageTitle{Waypoint}}
\description{\packageDescription{Waypoint}}
\section{Package Content}{\packageIndices{Waypoint}}
\author{\packageAuthor{Waypoint}}
\section{Maintainer}{\packageMaintainer{Waypoint}}
\keyword{package}
