% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoordBase.R, R/RcppExports.R
\name{Coords}
\alias{Coords}
\alias{as_coords}
\alias{print.coords}
\alias{as_coords.default}
\alias{format.coords}
\alias{as_coords.waypoints}
\title{Geographic or GPS Coordinate Class}
\usage{
as_coords(object, ...)

\method{print}{coords}(x, ...)

\method{as_coords}{default}(object, fmt = 1L)

\method{format}{coords}(x, usenames = TRUE)

\method{as_coords}{waypoints}(object, latlon)
}
\arguments{
\item{object}{a \code{numeric} vector of coordinate values, optionally named, or an object of
class \code{"waypoints"}.}

\item{\dots}{further arguments passed to or from other methods.}

\item{x}{object of class \code{"coords"} created by function
\code{\link[=as_coords]{as_coords}()}.}

\item{fmt}{\code{integer}, 1L, 2L or 3L, indicating the current or desired coordinate format.}

\item{usenames}{\code{logical}, whether or not to include names in formatted output.}

\item{latlon}{\code{logical}, indicating whether the \code{as_coords()} S3 method for class
\code{"waypoints"} extracts the latitude component of argument \code{object} (if \code{TRUE}),
or the longitude (if \code{FALSE}).}
}
\value{
An object of class \code{"coords"}, comprising the original a \code{numeric} vector argument
\code{nv} with values possibly converted as appropriate and additional attributes: –
\item{\code{"fmt"}}{the coordinate format.}
\item{\code{"valid"}}{a \code{logical} vector indicating whether individual coordinate values
  are valid geographic locations.}
}
\description{
\code{as_coords()} creates an object of class \code{"coords"}, a robust representation of a
series of geographic or GPS coordinate values.

\code{convert()} converts the format of existing objects of class \code{"coords"} between (i)
decimal degrees, (ii) degrees and minutes, and (iii) degrees, minutes and seconds.
}
\details{
Individual values provided in the \code{numeric} vector argument \code{nv} should have a decimal
point after the number of whole degrees in the case of \emph{decimal degrees}, after the number
of whole minutes in the case of \emph{degrees and minutes}, and after the number of whole
seconds in the case of \emph{degrees, minutes and seconds}.

The \code{fmt} argument is used to specify the format of values in the \code{numeric} vector
argument \code{object} to be converted into a \code{"coords"} object.

The values of a newly created \code{"coords"} object are checked to ensure they are valid
geographic locations as described under \code{\link[=validate]{validate}()}.
}
\examples{
## Numeric vector representing degrees and minutes
dm <- c(5130.4659, 4932.7726, 4806.4339, 3853.3696, 0.0000, -3706.7044, -5306.2869, -2514.4093,
        -007.6754, 1823.9137, -12246.7203, -7702.1145, 0.0000, -1217.3178, 7331.0370, -5731.1536)

## Create a "coords" object of degrees and minutes (fmt = 2)
as_coords(dm, fmt = 2)

## Name the "coords" object
names(dm) <- rep(c("Nelson's Column", "Ostravice", "Tally Ho", "Washington Monument", "Null Island",
                   "Tristan da Cunha", "Mawson Peak", "Silvio Pettirossi International Airport"), 2)
dm

## Decimal degrees as an ordinary R numeric vector
as.numeric(dm)

## Format as a fixed-width character vector witht names...
format(dm)

## ...or without them
format(dm, usenames = FALSE)

rm(dm)


}
\seealso{
\code{\link[base:attr]{attr}()}, \code{\link[base:attributes]{attributes}},
  \code{\link[=latlon]{latlon}()}, \code{\link[base:numeric]{numeric}()} and
  \code{\link[=validate]{validate}()}.

Other coords_waypoints: 
\code{\link{Waypoints}}
}
\concept{coords_waypoints}
