% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoordBase.R, R/RcppExports.R
\name{Coords}
\alias{Coords}
\alias{as_coords}
\alias{convert}
\alias{print.coords}
\alias{as_coords.default}
\alias{convert.coords}
\alias{format.coords}
\alias{as_coords.waypoints}
\title{Geographic or GPS Coordinate Class}
\usage{
as_coords(object, ...)

convert(x, ...)

\method{print}{coords}(x, ...)

\method{as_coords}{default}(object, fmt = 1L)

\method{convert}{coords}(x, fmt)

\method{format}{coords}(x, usenames = TRUE)

\method{as_coords}{waypoints}(object, latlon)
}
\arguments{
\item{object}{a \code{numeric} vector of coordinate values, optionally named, or an object of
class \code{"waypoints"}.}

\item{\dots}{further arguments passed to or from other methods.}

\item{x}{object of class \code{"coords"} created by function
\code{\link[=as_coords]{as_coords}()}.}

\item{fmt}{\code{integer}, 1L, 2L or 3L, indicating the current or desired coordinate format.}

\item{usenames}{\code{logical}, whether or not to include names in formatted output.}

\item{latlon}{\code{logical}, indicating whether the \code{as_coords()} S3 method for class
\code{"waypoints"} extracts the latitude component of argument \code{object} (if \code{TRUE}),
or the longitude (if \code{FALSE}).}
}
\value{
An object of class \code{"coords"}, comprising the original a \code{numeric} vector argument
\code{nv} with values possibly converted as appropriate and additional attributes: –
\item{\code{"fmt"}}{the coordinate format.}
\item{\code{"valid"}}{a \code{logical} vector indicating whether individual coordinate values
  are valid geographic locations.}
}
\description{
\code{as_coords()} creates a robust representation of a series of geographic or GPS
coordinates instantiated as an object of class \code{"coords"}.

\code{convert()} converts the format of existing objects of class \code{"coords"} between (i)
decimal degrees, (ii) degrees and minutes, and (iii) degrees, minutes and seconds.
}
\details{
Individual values provided in the \code{numeric} vector argument \code{nv} should have a decimal
point after the number of whole degrees in the case of \emph{decimal degrees}, after the number
of whole minutes in the case of \emph{degrees and minutes}, and after the number of whole
seconds in the case of \emph{degrees, minutes and seconds}.

The \code{fmt} argument should be \code{1L} to represent decimal degrees, \code{2L} for degrees
and minutes, and \code{3L} for degrees, minutes and seconds and is used to provide both the
format of values in \code{numeric} vector argument \code{nv} to be converted into a
\code{"coords"} object and the desired format if a \code{"coords"} object is to be converted to
a new format. Note that on conversion of a \code{"coords"} object, the original \code{numeric}
vector argument \code{nv} is modified such that the values are as described in the previous
paragraph, and may be inspected using standard R code, see examples.

The values of a newly created \code{"coords"} object are checked to ensure they are valid
geographic locations as described under \code{\link[=validate]{validate}()}. Likewise, a
check is made to ensure that an existing \code{"coords"} object to be converted to a new format
has already been validated; if not, it is re-validated.
}
\examples{
## Numeric vector representing degrees and minutes
dm <- c(5130.4659, 4932.7726, 4806.4339, 3853.3696, 0.0000, -3706.7044, -5306.2869, -2514.4093,
        -007.6754, 1823.9137, -12246.7203, -7702.1145, 0.0000, -1217.3178, 7331.0370, -5731.1536)

## Create a "coords" object of degrees and minutes (fmt = 2)
as_coords(dm, fmt = 2)

## Name the "coords" object
names(dm) <- rep(c("Nelson's Column", "Ostravice", "Tally Ho", "Washington Monument", "Null Island",
                   "Tristan da Cunha", "Mawson Peak", "Silvio Pettirossi International Airport"), 2)
dm

## Convert to degrees, minutes and seconds (fmt = 3)
convert(dm, 3)

## Convert to decimal degrees (fmt = 1)
convert(dm, 1)

## Decimal degrees as an ordinary R numeric vector
as.numeric(dm)

## Convert to degrees and minutes, then format as a fixed-width
## character vector without names…
convert(dm, 3) |> format(usenames = FALSE)

## …or with them
format(dm)

rm(dm)


}
\seealso{
\code{\link[base:attr]{attr}()}, \code{\link[base:attributes]{attributes}},
  \code{\link[=latlon]{latlon}()}, \code{\link[base:numeric]{numeric}()} and
  \code{\link[=validate]{validate}()}.

Other coords_waypoints: 
\code{\link{print.waypoints}()}
}
\concept{coords_waypoints}
