% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoordBase.R
\name{op-zero-length}
\alias{op-zero-length}
\alias{\%L\%}
\title{Operator Providing Alternative to Zero-Length Object}
\usage{
x \%L\% y
}
\arguments{
\item{x, y}{atomic vector arguments or other objects for which `length()` is defined.}
}
\value{
`x`, or if `length(x)` is zero, `y`.
}
\description{
Infix function implementing provision of an alternative if an object has zero length.
}
\details{
The infix function `%L%` may be useful in implementing `if (length(x)) x else y` and was inspired by
the null coalescing operator [`%||%`][base::Control].
}
\examples{
c1 <- letters[1:4]
c2 <- character(0)
n1 <- 1:3
n2 <- numeric(0)

c1 \%L\% n1
c2 \%L\% n1

n1 \%L\% c1
n2 \%L\% c1

rm(c1, c2, n1, n2)
}
\seealso{
[`%||%`][base::Control].
}
\concept{utils}
\keyword{arith}
\keyword{univar}
