% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoordBase.R
\name{review}
\alias{review}
\alias{review.coords}
\title{Review Coordinates and Waypoints}
\usage{
review(x, ...)

\method{review}{coords}(x, ..., show_n = 20L)
}
\arguments{
\item{x}{object of class \code{"coords"} or \code{"waypoints"}.}

\item{...}{further arguments passed to or from other methods.}

\item{show_n}{\code{integer}, the maximum number of invalid elements of argument \code{x} to
include in the output; default \code{20L}.}
}
\value{
A \code{list} comprising the following components: -

\item{allvalid}{\code{logical}, whether or not all the elements of argument \code{x} are valid.}
\item{n_invalid}{\code{integer}, the number of invalid elements in argument \code{x}, if any.}
\item{invalids}{\code{numeric} vector including invalid elements of argument \code{x}, if any.}
\item{which_invalid}{\code{integer} vector specifying which elements of argument \code{x} are
	 invalid, if any.}
}
\description{
Review validity of elements of \code{"coords"} and  \code{"waypoints"} objects
}
\details{
\code{review()} reveals elements of \code{"coords"} and  \code{"waypoints"} objects that do not
conform to the criteria checked by \code{\link[=validate]{validate}()}, i.e. are not valid
geographic locations.
}
\seealso{
\code{\link[=coords]{"coords"}} and \code{\link[=waypoints]{"waypoints"}}.

Other validate: 
\code{\link{validate}()}
}
\concept{validate}
