% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoordBase.R, R/RcppExports.R
\name{review}
\alias{review}
\alias{review.coords}
\alias{review.waypoints}
\alias{as_coords}
\title{Review Coordinates and Waypoints Validity}
\usage{
review(x, ...)

\method{review}{coords}(x, show_n = 20L, ...)

\method{review}{waypoints}(x, show_n = 20L, ...)

as_coords(wp, latlon)
}
\arguments{
\item{x}{object of class \code{"coords"} or \code{"waypoints"}.}

\item{\dots}{further arguments passed to or from other methods.}

\item{show_n}{\code{integer}, the maximum number of invalid elements of argument \code{x} to
include in the output; default \code{20L}.}

\item{wp}{object of class \code{"waypoints"}.}

\item{latlon}{logical indicating whether \code{as_coords()} extracts the latitude component of
argument \code{wp} if \code{TRUE}, or if \code{FALSE} the longitude.}
}
\value{
The \code{review()} method for class \code{"coords"} returns a \code{\link[base:list]{list}} comprising the
following elements: -

\item{allvalid}{\code{logical}, whether or not all the elements of argument \code{x} are valid.}
\item{n_invalid}{\code{integer}, the number of invalid elements in argument \code{x}, if any.}
\item{invalids}{\code{numeric} vector including invalid elements of argument \code{x}, if any.}
\item{which_invalid}{\code{integer} vector specifying which elements of argument \code{x} are
	 invalid, if any.}

The method for class \code{"waypoints"} returns a list of two sub-lists, each sub-list with
elements as described for the method for class \code{"coords"}, one each for latitude and
longitude.

\code{as_coord} returns an object of class \code{"coords"} based on either the latitude or longitude.
}
\description{
\code{review()} review validity of elements of \code{"coords"} and \code{"waypoints"} objects.

\code{as_coords()} creates an object of class \code{"coords"} based on either the latitude or
longitude components of a \code{"waypoints"} object.
}
\details{
\code{review()} reveals elements of \code{"coords"} and  \code{"waypoints"} objects that do not
conform to the criteria checked by \code{\link[=validate]{validate}()}, i.e. are not valid
geographic locations.

\code{as_coord()} is primarily intended for internal use by \code{review()} but is also exposed
to the user as potentially of use.
}
\examples{
## Continuing example from `validate()`, named numeric vector representing degrees and minutes,
## the erroneous first value having more than 60 minutes
\dontshow{
   oldopt <- options(warn = -1)
   dm <-
       c(5160.4659, 4932.7726, 4806.4339, 3853.3696, 0.0000, -3706.7044, -5306.2869, -2514.4093,
	   -007.6754, 1823.9137, -12246.7203, -7702.1145, 0.0000, -1217.3178, 7331.0370, -5731.1536)

   names(dm) <- 
       rep(c("Nelson's Column", "Ostravice", "Tally Ho", "Washington Monument", "Null Island",
             "Tristan da Cunha", "Mawson Peak", "Silvio Pettirossi International Airport"), 2)
}

## Create "coords" object of degrees and minutes
coords(dm) <- 2

review(dm)

###
## Continuing example from `validate()`, data frame representing waypoint names and latitude
## and longitude values in decimal degrees, the erroneous penultimate latitude having more than
## 90 degrees absolute value
\dontshow{
wp1 <- data.frame(
    name = c("Nelson's Column", "Ostravice", "Tally Ho", "Washington Monument", "Null Island",
             "Tristan da Cunha", "Mawson Peak", "Silvio Pettirossi International Airport"),
    lat = c(51.507765, 49.54621, 48.107232, 38.889494, 0, -37.11174, -93.104781, -25.240156),
    lon = c(-0.127924, 18.398562, -122.778671, -77.035242, 0, -12.28863, 73.517283, -57.519227)
)
}

## Create "waypoints" object of decimal degrees
waypoints(wp1) <- 1

review(wp1)

rm(dm, wp1)

\dontshow{
   options(oldopt)
}
}
\seealso{
\code{"\link[=coords]{coords}"} and \code{"\link[=waypoints]{waypoints}"}.

Other validate: 
\code{\link{validate}()}
}
\concept{validate}
