% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{validate}
\alias{validate}
\alias{validate.coords}
\alias{validate.waypoints}
\title{Validate Coords or Waypoints}
\usage{
\method{validate}{coords}(nv)

\method{validate}{waypoints}(df)
}
\arguments{
\item{nv}{object of class \code{"coords"}.}

\item{df}{object of class \code{"waypoints"}.}
}
\value{
\code{validate()} returns its argument with \code{boolean vector} attribute \code{"valid"}
updated as appropriate.
}
\description{
\code{validate()} validate objects of class \code{"coords"} or \code{"waypoints"}.
}
\details{
Individual coordinate values within \code{\link[=coords]{"coords"}} or
\code{\link[=waypoints]{"waypoints"}} objects are validated to ensure their being plausible
geographic locations.

The values of ...
}
