% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoordBase.R, R/RcppExports.R
\name{validate}
\alias{validate}
\alias{validate.coords}
\alias{validate.waypoints}
\title{Validate Coords or Waypoints}
\usage{
validate(x, ...)

\method{validate}{coords}(x)

\method{validate}{waypoints}(x)
}
\arguments{
\item{x}{object of class \code{"\link{coords}"} created by function \code{\link{as_coords}()},
or of class \code{"\link{waypoints}"} created by function \code{\link{as_waypoints}()}.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
\code{validate()} returns its argument with \code{logical} vector attribute \code{"valid"},
or attributes \code{"validlat"} and \code{"validlon"} updated as appropriate for
\code{"coords"} and' \code{"waypoints"} objects respectively.
}
\description{
\code{validate()} validates objects of class \code{"coords"} or \code{"waypoints"}.
}
\details{
Individual coordinate values within \code{"\link{coords}"} or
\code{"\link{waypoints}"} objects are checked to ensure they represent valid
geographic locations.

To be valid, the absolute values of coordinates in degrees must not exceed 180, or 90 if degrees
of latitude and, similarly, the absolute values of the minutes and seconds components, where
given, must not exceed 60 degrees. Otherwise a warning will be issued and the \code{"valid"}
attribute in the case of a \code{"coords"} object, or \code{"validlat"} and \code{"validlon"}
attributes in the case of a \code{"waypoints"} object will be set to \code{FALSE} for any
non-compliant coordinate values.
}
\examples{
## Continuing example from `coords()`...
## Create "coords" object in degrees and minutes
\dontshow{
   dm <-
       c(5130.4659, 4932.7726, 4806.4339, 3853.3696, 0.0000, -3706.7044, -5306.2869, -2514.4093,
	   -007.6754, 1823.9137, -12246.7203, -7702.1145, 0.0000, -1217.3178, 7331.0370, -5731.1536)

   names(dm) <- 
       rep(c("Nelson's Column", "Ostravice", "Tally Ho", "Washington Monument", "Null Island",
             "Tristan da Cunha", "Mawson Peak", "Silvio Pettirossi International Airport"), 2)

as_coords(dm, fmt = 2)
}

validate(dm)

## Change first value to have more than 60 minutes
dm[1] <- 5160.4659

validate(dm)

## Examine "valid" attribute of dm
attr(dm, "valid")

###
## Continuing example from `waypoints()`...
## Create "waypoints" object in decimal degrees
\dontshow{
wp1 <- data.frame(
    name = c("Nelson's Column", "Ostravice", "Tally Ho", "Washington Monument", "Null Island",
             "Tristan da Cunha", "Mawson Peak", "Silvio Pettirossi International Airport"),
    lat = c(51.507765, 49.54621, 48.107232, 38.889494, 0, -37.11174, -53.104781, -25.240156),
    lon = c(-0.127924, 18.398562, -122.778671, -77.035242, 0, -12.28863, 73.517283, -57.519227)
)

as_waypoints(wp1, fmt = 1)
}

validate(wp1)

## Change penultimate latitude absolute value to have more than 90 degrees
wp1$lat[7] <- -93.104781

validate(wp1)

## Examine "validlat" attribute of wp1
attr(wp1, "validlat")

rm(dm, wp1)

}
\seealso{
\code{"\link{coords}"} and \code{"\link{waypoints}"}.

Other validate: 
\code{\link{review}()}
}
\concept{validate}
