% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{waypoints}
\alias{waypoints}
\alias{waypoints<-}
\alias{print.waypoints}
\title{Geographic or GPS Waypoint Class}
\usage{
waypoints(df, fmt = 1L)

waypoints(df) <- value

\method{print}{waypoints}(wp)
}
\arguments{
\item{df}{a data frame with each row representing a waypoint, comprising at least two
\code{numeric} columns containing values of latitude and longitude, and optionally a
\code{character} column of waypoint names (see \emph{Details}).}

\item{fmt, value}{an \code{integer} of value 1L, 2L or 3L, indicating the current or desired
coordinate format (see \emph{Details}); default 1L.}

\item{wp}{an object of class \code{"waypoints"} created by function
\code{\link[=waypoints]{waypoints}()}.}
}
\value{
An object of classes \code{"waypoints"} and \code{"data.frame"}, comprising the original data
frame argument \code{df}, with latitude and longitude values possibly converted as appropriate
and additional attributes: –
\item{\code{"fmt"}}{the coordinate format.}
\item{\code{"namescol"}}{the position of waypoint names, if present within \code{df}.}
\item{\code{"llcols"}}{the position of latitude and longitude columns within \code{df}.}
\item{\code{"validlat"} and \code{"validlon"}}{\code{logical} vectors indicating whether
  individual latitude and longitude values are valid geographic locations.}
}
\description{
\code{waypoints()} creates a robust representation of a series of geographic or GPS waypoints
instantiated as an object of class \code{"waypoints"}.

\code{waypoints()} and replacement form \verb{waypoints()<-} also convert the format of
existing objects of class \code{"waypoints"} between (i) decimal degrees, (ii) degrees and
minutes, and (iii) degrees, minutes and seconds.
}
\details{
By default, the names of the waypoints should be included in a "Name" column of data frame
argument \code{df}, and the latitude and longitude in the two columns immediately on the right
hand side of "Name". An alternative column for waypoint names may be specified by setting an
\code{integer} attribute, \code{"namescol"} indicating its position in \code{df}, while setting
this attribute to \code{NA} supresses printing of waypoint names. If \code{df} has neither a
"Name" column nor a \code{"namescol"} attribute, the \code{"row.names"} attribute is used for
waypoint names if present in \code{df}. Similarly, alternative columns for the latitude and
longitude may be specified by setting \code{"llcols"} as a length 2 \code{integer} vector
attribute indicating their positions in \code{df}.

Individual values provided in the \code{numeric} vector latitude and longitude columns of data
frame argument \code{df} should have a decimal point after the number of whole degrees in the
case of \emph{decimal degrees}, after the number of whole minutes in the case of
\emph{degrees and minutes}, and after the number of whole seconds in the case of
\emph{degrees, minutes and seconds}.

The \code{fmt} argument should be \code{1L} to represent decimal degrees, \code{2L} for degrees
and minutes, and \code{3L} for degrees, minutes and seconds and is used to provide both the
format of values in data frame argument \code{df} to be converted into a \code{"waypoints"}
object and the desired format if a \code{"waypoints"} object is to be converted to a new format.
Note that on conversion of a \code{"waypoints"} object, the original data frame argument
\code{df} is modified such that the latitude and longitude values are as described in the
previous paragraph, and may be inspected using standard R code, see examples.

The latitude and longitude values of a newly created \code{"waypoints"} object are checked to
ensure they are valid geographic locations as described under
\code{\link[=validate]{validate}()}. Likewise, a check is made to ensure that an existing
\code{"waypoints"} object to be converted to a new format has already been validated; if not, it
is re-validated.
}
\examples{
## Dataframe representing waypoint names and latitude and longitude values in degrees, minutes and seconds
wp1 <- data.frame(
    name = c("Nelson's Column", "Ostravice", "Tally Ho", "Washington Monument", "Null Island",
             "Tristan da Cunha", "Mawson Peak", "Silvio Pettirossi International Airport"),
    lat = c(513027.95, 493246.36, 480626.04, 385322.18, 00000.00, -370642.26, -530617.21, -573109.21),
    lon = c(-00740.53, 182354.82, -1224643.22, -770206.87, 00000.00, -121719.07, 733102.22, -251424.56)
)

## Create "waypoints" object of degrees, minutes and seconds (fmt = 3)
waypoints(wp1, 3)

## Convert to degrees and minutes (fmt = 2)
waypoints(wp1) <- 2
wp1

## Convert to decimal degrees (fmt = 1)
waypoints(wp1) <- 1
wp1

###
## Dataframe representing unnamed latitude and longitude values in decimal degrees
wp2 <- data.frame(
    lat = c(51.507765, 49.54621, 48.107232, 38.889494, 0, -37.11174, -53.104781, -57.519227),
    lon = c(-0.127924, 18.398562, -122.778671, -77.035242, 0, -12.28863, 73.517283, -25.240156)
)

## Create "waypoints" object of decimal degrees (default fmt = 1)
waypoints(wp2)

## Add row.names
row.names(wp2) <-
    c("Nelson's Column", "Ostravice", "Tally Ho", "Washington Monument", "Null Island",
      "Tristan da Cunha", "Mawson Peak", "Silvio Pettirossi International Airport")
wp2

## Convert to degrees and minutes (fmt = 2)
waypoints(wp2) <- 2
wp2

## Convert to degrees, minutes and seconds (fmt = 3)
waypoints(wp2) <- 3
wp2

## Show degrees, minutes and seconds as a bare data frame
as.data.frame(wp2)

rm(wp1, wp2)

}
\seealso{
\code{\link[base:attr]{attr()}}, \code{\link[base:attributes]{attributes}},
  \code{\link[base:data.frame]{data.frame()}}, and \code{\link[=validate]{validate}()}.

Other coords_waypoints: 
\code{\link{coords}()}
}
\concept{coords_waypoints}
